+++ gradle.properties
@@ -1,3 +1,4 @@
-projectVersion=3.1.2-SNAPSHOT
+projectVersion=3.2.0-SNAPSHOT
 githubBranch=master
 org.gradle.caching=true
+
