+++ src/main/java/io/micronaut/gradle/MicronautMinimalLibraryPlugin.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2003-2021 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micronaut.gradle;
+
+import org.gradle.api.Plugin;
+import org.gradle.api.Project;
+import org.gradle.api.plugins.JavaLibraryPlugin;
+import org.gradle.api.plugins.PluginManager;
+
+/**
+ * A plugin to build a Micronaut library. It is the base plugin for all Micronaut libraries,
+ * and applies the _minimal_ amount of plugins to make it work.
+ */
+public class MicronautMinimalLibraryPlugin implements Plugin<Project> {
+    @Override
+    public void apply(Project project) {
+        PluginManager plugins = project.getPluginManager();
+        plugins.apply(JavaLibraryPlugin.class);
+        plugins.apply(MicronautComponentPlugin.class);
+    }
+}
